sr: move post pole emploi user controller from config to route
